requirements.txt
# requirements.txt  – ScalperBot 2.0 minimal runtime stack
pandas>=2.2.2
numpy>=1.26.4
ta>=0.11.0              # technical-analysis indicators (builds on pandas & numpy)
python-dateutil>=2.9.0
six>=1.16.0             # transient dep of python-dateutil but pinned to avoid misses
python-dotenv>=1.0.1
openai>=1.25.0
coinbase-advanced-py>=0.1.5
psutil>=5.9.8           # system monitoring for production runner
websocket-client>=1.7.0 # WebSocket feed connection
requests>=2.31.0        # HTTP requests with retry logic
tenacity>=8.2.3         # Retry decorator for API calls
scipy>=1.11.0           # Scientific computing for advanced analytics
scikit-learn>=1.3.0     # Machine learning for ML predictor
joblib>=1.3.0           # ML model serialization
PyYAML>=6.0             # YAML configuration for dynamic config
colorama>=0.4.6         # Colored terminal output
pandas-ta>=0.3.14b0     # Advanced technical indicators
